refactor: split main process entrypoint
Closes #115
