+++ src/main/app-state.ts
@@ -1,0 +1,9 @@
+let isQuitting = false
+
+export function getIsQuitting() {
+  return isQuitting
+}
+
+export function setIsQuitting(value = true) {
+  isQuitting = value
+}
